+++ feedbacker/src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod feedback;
+mod index;
+
+pub use feedback::*;
+pub use index::*;
